done all features

+++ src/main/java/org/example/controller/StudentController.java
@@ -1,7 +1,10 @@
 package org.example.controller;
 
 import org.example.model.Student;
+import org.example.model.UserDetails;
+import org.example.model.UserPermission;
 import org.example.repository.StudentRepository;
+import org.example.security.AuthService;
 
 import java.sql.SQLException;
 import java.util.List;
@@ -13,20 +16,42 @@ public class StudentController {
         repository = new StudentRepository();
     }
 
-    public List<Student> loadStudents() {
-        return repository.getAllStudents();
+    public List<Student> loadStudents(UserDetails user) {
+        if (AuthService.hasPermission(UserPermission.READ, user))
+        {
+            return repository.getAllStudents();
+        } else {
+            throw new SecurityException("Permission denied");
+        }
+
     }
 
-    public void createStudent(Student student) throws SQLException {
-        repository.addStudent(student);
+    public void createStudent(UserDetails user, Student student) throws SQLException {
+        if (AuthService.hasPermission(UserPermission.WRITE, user))
+        {
+            repository.addStudent(student);
+        } else {
+            throw new SecurityException("Permission denied");
+        }
     }
 
-    public void updateStudent(Student student) throws SQLException {
-        repository.updateStudent(student);
+    public void updateStudent(UserDetails user, Student student) throws SQLException {
+        if (AuthService.hasPermission(UserPermission.UPDATE, user))
+        {
+            repository.updateStudent(student);
+        } else {
+            throw new SecurityException("Permission denied");
+        }
     }
 
-    public void deleteStudent(int studentId) throws SQLException {
-        repository.deleteStudent(studentId);
+    public void deleteStudent(UserDetails user, int studentId) throws SQLException {
+
+        if (AuthService.hasPermission(UserPermission.DELETE, user))
+        {
+            repository.deleteStudent(studentId);
+        } else {
+            throw new SecurityException("Permission denied");
+        }
     }
 
 }
